refactor(site/src/pages/AgentsPage): remove useThinkingDisplayMode hook, use Radix Collapsible
