+++ site/src/pages/AgentsPage/components/ThinkingDisplaySettings.tsx
@@ -1,6 +1,10 @@
 import type { FC } from "react";
+import { useMutation, useQuery, useQueryClient } from "react-query";
+import {
+	preferenceSettings,
+	updatePreferenceSettings,
+} from "#/api/queries/users";
 import type { ThinkingDisplayMode } from "#/api/typesGenerated";
-import { useThinkingDisplayMode } from "../hooks/useThinkingDisplayMode";
 
 const options: { value: ThinkingDisplayMode; label: string }[] = [
 	{ value: "auto", label: "Auto" },
@@ -10,7 +14,11 @@ const options: { value: ThinkingDisplayMode; label: string }[] = [
 ];
 
 export const ThinkingDisplaySettings: FC = () => {
-	const { mode, setMode } = useThinkingDisplayMode();
+	const queryClient = useQueryClient();
+	const query = useQuery(preferenceSettings());
+	const mutation = useMutation(updatePreferenceSettings(queryClient));
+
+	const mode: ThinkingDisplayMode = query.data?.thinking_display_mode || "auto";
 
 	return (
 		<div className="flex flex-col gap-2">
@@ -26,7 +34,14 @@ export const ThinkingDisplaySettings: FC = () => {
 				</p>
 				<select
 					value={mode}
-					onChange={(e) => setMode(e.target.value as ThinkingDisplayMode)}
+					onChange={(e) =>
+						mutation.mutate({
+							...query.data,
+							thinking_display_mode: e.target.value as ThinkingDisplayMode,
+							task_notification_alert_dismissed:
+								query.data?.task_notification_alert_dismissed ?? false,
+						})
+					}
 					aria-label="Thinking display mode"
 					className="rounded-md border border-border bg-surface-primary px-2 py-1 text-xs text-content-primary"
 				>
